Make changes to Gemfile before starting "production"
This commit updates the 'Gemfile' and 'Gemfile.lock' files in order to
use the postgreSQL for heroku. The 'sqlite3' gem was moved to the
':development' and ':test' groups and the ':production' group was added
with only the 'pg' gem.

The 'Gemfile.lock' was updated automatically by executing
'bundle install --without production'